+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   **BREAKING CHANGES**
   - [#156](https://github.com/Datatamer/unify-client-python/issues/156) Fetch Dataset profile, even if out of date.
   - [#161](https://github.com/Datatamer/unify-client-python/issues/161) Move `create_attribute` from Dataset to AttributeCollection 
+  - The `Project` method `add_source_dataset` has been renamed `add_input_dataset` to model the API endpoint.
 
   **NEW FEATURES**
   - [#65](https://github.com/Datatamer/unify-client-python/issues/65) Fetches published clusters with data represented as a dataset.
@@ -10,6 +11,7 @@
   - [#113](https://github.com/Datatamer/unify-client-python/issues/113) Add support for uploading a binningModel
   - [#168](https://github.com/Datatamer/unify-client-python/issues/168) Add support for project attributes
   - [#171](https://github.com/Datatamer/unify-client-python/issues/171) Add support for creating and retrieving taxonomies
+  - [#178](https://github.com/Datatamer/unify-client-python/issues/178) Add support for retrieving input datasets
 
   **BUG FIXES**
   - [#148](https://github.com/Datatamer/unify-client-python/issues/148) Fix null geo, id and absent id bug for geospatial datasets.